Refactored and added tests related to #238
* Minor refactoring and added tests related to #238 Cardinality
  validation support
