+++ src/Spark.Engine.Test/Extensions/OperationOutcomeExtensionsTests.cs
@@ -2,8 +2,6 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Spark.Engine.Extensions;
 using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace Spark.Engine.Test.Extensions
 {
@@ -11,26 +9,66 @@ namespace Spark.Engine.Test.Extensions
     public class OperationOutcomeExtensionsTests
     {
         [TestMethod]
-        public void CanConvertFhirPathExpressionToXPathExpression()
+        public void Three_Level_InnerErrors_Test()
         {
-            var fhirPathExpression = "Patient.name[0].family";
-            var expected = "f:Patient/f:name[0]/f:family";
+            OperationOutcome outcome;
 
-            var actual = OperationOutcomeExtensions.ConvertToXPathExpression(fhirPathExpression);
+            try
+            {
+                try
+                {
+                    try
+                    {
+                        throw new Exception("Third error level");
+                    }
+                    catch (Exception e3)
+                    {
+                        throw new Exception("Second error level", e3);
+                    }
+                }
+                catch (Exception e2)
+                {
+                    throw new Exception("First error level", e2);
+                }
+            }
+            catch (Exception e1)
+            {
+                outcome = new OperationOutcome().AddAllInnerErrors(e1);
+            }
 
-            Assert.AreEqual(expected, actual);
+            Assert.IsTrue(outcome.Issue.FindIndex(i => i.Diagnostics.Equals("Exception: First error level")) == 0, "First error level should be at index 0");
+            Assert.IsTrue(outcome.Issue.FindIndex(i => i.Diagnostics.Equals("Exception: Second error level")) == 1, "Second error level should be at index 1");
+            Assert.IsTrue(outcome.Issue.FindIndex(i => i.Diagnostics.Equals("Exception: Third error level")) == 2, "Third error level should be at index 2");
         }
 
         [TestMethod]
-        public void CanResolveToFhirPathExpression()
+        public void IssueSeverity_Is_Information_When_HttpStatusCode_Is_Continue_Test()
         {
-            var resourceType = typeof(Patient);
-            var expression = "Name[0].FamilyElement";
-            var expected = "Patient.name[0].family";
+            Assert.AreEqual(OperationOutcome.IssueSeverity.Information, OperationOutcomeExtensions.IssueSeverityOf(System.Net.HttpStatusCode.Continue));
+        }
 
-            var actual = OperationOutcomeExtensions.ResolveToFhirPathExpression(resourceType, expression);
+        [TestMethod]
+        public void IssueSeverity_Is_Information_When_HttpStatusCode_Is_Created_Test()
+        {
+            Assert.AreEqual(OperationOutcome.IssueSeverity.Information, OperationOutcomeExtensions.IssueSeverityOf(System.Net.HttpStatusCode.Created));
+        }
 
-            Assert.AreEqual(expected, actual);
+        [TestMethod]
+        public void IssueSeverity_Is_Warning_When_HttpStatusCode_Is_MovedPermanently_Test()
+        {
+            Assert.AreEqual(OperationOutcome.IssueSeverity.Warning, OperationOutcomeExtensions.IssueSeverityOf(System.Net.HttpStatusCode.MovedPermanently));
+        }
+
+        [TestMethod]
+        public void IssueSeverity_Is_Error_When_HttpStatusCode_Is_BadRequest_Test()
+        {
+            Assert.AreEqual(OperationOutcome.IssueSeverity.Error, OperationOutcomeExtensions.IssueSeverityOf(System.Net.HttpStatusCode.BadRequest));
+        }
+
+        [TestMethod]
+        public void IssueSeverity_Is_Fatal_When_HttpStatusCode_Is_InternalServerError_Test()
+        {
+            Assert.AreEqual(OperationOutcome.IssueSeverity.Fatal, OperationOutcomeExtensions.IssueSeverityOf(System.Net.HttpStatusCode.InternalServerError));
         }
     }
 }
